+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Develop
+- Add: date_from in get_all_entries
 - Add-Config: POSTGRES_DB_POOL_SIZE and POSTGRES_DB_MAX_OVERFLOW
 
 ## V0.1.1